Put Type based field ignoring classes into ignore package.
